refactoring and integration of latest changes
NOT BUGFREE
- added own uncurry3 method in own module Tuple.hs
- exported most game functionality into own module Game.hs
- added handling of fire input, unfortunately does not fully work yet - game crashes on fire button press - further investigation necessary
